small update of catalog.,cpp

+++ catalog.cpp
@@ -93,7 +93,7 @@ void Catalog::sortbyDate() {
 
 // Method to print the catalog
 void Catalog::printCatalog() const {
-    for (size_t i = 0; i < library.size(); ++i) {
+    for (int i = 0; i < library.size(); ++i) {
         library[i].printBook();
     }
     return false;
